Feature/165415456 location filter

+++ app/controllers/api/events_controller.rb
@@ -102,7 +102,7 @@ class API::EventsController < API::BaseController
 
     end
 
-    miles = miles.to_i
+    miles = miles.to_f
     if miles.positive?
       location = []
       location = [latitude, longitude] if [latitude, longitude].all?(&:present?)
